requirements.txt
aniso8601==8.0.0
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cython
dependency-injector==3.12.4
Flask==1.0.2
Flask-RESTful==0.3.6
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
joblib==0.13.2
jsonschema==3.0.2
MarkupSafe==1.1.1
msgpack==0.6.1
numpy==1.15.0
pandas=0.23.3
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.7.1
pytz==2019.2
pyzmq==18.1.0
requests==2.22.0
scikit-learn==0.21.3
simplejson==3.16.0
six==1.11.0
urllib3==1.25.3
Werkzeug==0.15.6
